refactor(faunafinder): separate resolving the organization from when it runs
The resolver implemented IHostedService, so it decided its own lifecycle and
every caller inherited that choice. It is now a plain scoped service exposing
ResolveAsync, and FaunaFinderOrganizationStartupTask owns the decision to run
it once before serving traffic.

Dropping IHostedService also drops IServiceScopeFactory from the resolver: a
scoped service injects IOrganizationRepository directly, and only the startup
task needs a scope. Moving to lazy resolution or a periodic refresh now means
replacing the startup task alone.
